requirements.txt
python-dotenv
apache-airflow
pandas
numpy
mlflow
flake8
skops

-e .
